# docs/swagger.yaml
definitions:
  github_com_DIMO-Network_valuations-api_internal_core_models.DeviceOffer:
    properties:
      offerSets:
        description: Contains a list of offer sets, one for each source
        items:
          $ref: '#/definitions/github_com_DIMO-Network_valuations-api_internal_core_models.OfferSet'
        type: array
    type: object
  github_com_DIMO-Network_valuations-api_internal_core_models.DeviceValuation:
    properties:
      valuationSets:
        description: Contains a list of valuation sets, one for each vendor
        items:
          $ref: '#/definitions/github_com_DIMO-Network_valuations-api_internal_core_models.ValuationSet'
        type: array
    type: object
  github_com_DIMO-Network_valuations-api_internal_core_models.OdometerMeasurementEnum:
    enum:
    - Real
    - Estimated
    - Market
    type: string
    x-enum-varnames:
    - Real
    - Estimated
    - Market
  github_com_DIMO-Network_valuations-api_internal_core_models.Offer:
    properties:
      declineReason:
        description: The reason the offer was declined from the vendor
        type: string
      error:
        description: An error from the vendor (eg. when the VIN is invalid)
        type: string
      grade:
        description: The grade of the offer from the vendor (eg. "RETAIL")
        type: string
      price:
        description: The offer price from the vendor
        type: integer
      url:
        description: The offer URL from the vendor
        type: string
      vendor:
        description: The vendor of the offer (eg. "carmax", "carvana", etc.)
        type: string
    type: object
  github_com_DIMO-Network_valuations-api_internal_core_models.OfferSet:
    properties:
      mileage:
        description: The mileage used for the offers
        type: integer
      odometerMeasurementType:
        allOf:
        - $ref: '#/definitions/github_com_DIMO-Network_valuations-api_internal_core_models.OdometerMeasurementEnum'
        description: Whether estimated, real, or from the market (eg. vincario). Market,
          Estimated, Real
      offers:
        description: Contains a list of offers from the source
        items:
          $ref: '#/definitions/github_com_DIMO-Network_valuations-api_internal_core_models.Offer'
        type: array
      source:
        description: The source of the offers (eg. "drivly")
        type: string
      updated:
        description: The time the offers were pulled
        type: string
      zipCode:
        description: This will be the zip code used (if any) for the offers request
          regardless if the source uses it
        type: string
    type: object
  github_com_DIMO-Network_valuations-api_internal_core_models.ValuationSet:
    properties:
      currency:
        description: eg. USD or EUR
        type: string
      mileage:
        description: The mileage used for the valuation
        type: integer
      odometer:
        description: Odometer used for calculating values
        type: integer
      odometerMeasurementType:
        allOf:
        - $ref: '#/definitions/github_com_DIMO-Network_valuations-api_internal_core_models.OdometerMeasurementEnum'
        description: whether estimated, real, or from the market (eg. vincario). Market,
          Estimated, Real
      odometerUnit:
        type: string
      retail:
        description: retail is equal to retailAverage when available
        type: integer
      retailAverage:
        type: integer
      retailClean:
        description: retailClean, retailAverage, and retailRough my not always be
          available
        type: integer
      retailRough:
        type: integer
      retailSource:
        description: Useful when Drivly returns multiple vendors and we've selected
          one (eg. "drivly:blackbook")
        type: string
      tradeIn:
        description: tradeIn is equal to tradeInAverage when available
        type: integer
      tradeInAverage:
        type: integer
      tradeInClean:
        description: tradeInClean, tradeInAverage, and tradeInRough my not always
          be available
        type: integer
      tradeInRough:
        type: integer
      tradeInSource:
        description: Useful when Drivly returns multiple vendors and we've selected
          one (eg. "drivly:blackbook")
        type: string
      updated:
        description: The time the valuation was pulled or in the case of blackbook,
          this may be the event time of the device odometer which was used for the
          valuation
        type: string
      userDisplayPrice:
        description: UserDisplayPrice the top level value to show to users in mobile
          app
        type: integer
      vendor:
        description: The source of the valuation (eg. "drivly" or "blackbook")
        type: string
      zipCode:
        description: This will be the zip code used (if any) for the valuation request
          regardless if the vendor uses it
        type: string
    type: object
info:
  contact: {}
  description: API to get latest valuation for a given connected vehicle belonging
    to user. Tokens must be privilege tokens.
  title: DIMO Vehicle Valuations API
  version: "1.0"
paths:
  /v2/vehicles/{tokenId}/instant-offer:
    post:
      description: |-
        makes a request for an instant offer for a particular user device. Simply returns success if able to create job.
        You will need to query the offers endpoint to see if a new offer showed up. Job can take about a minute to complete.
      parameters:
      - description: tokenId for vehicle to get offers
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - offers
  /v2/vehicles/{tokenId}/offers:
    get:
      description: gets any existing offers for a particular user device. You must
        call instant-offer endpoint first to pull newer. Returns list.
      parameters:
      - description: tokenId for vehicle to get offers
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DIMO-Network_valuations-api_internal_core_models.DeviceOffer'
      security:
      - BearerAuth: []
      tags:
      - offers
  /v2/vehicles/{tokenId}/valuation:
    post:
      description: request valuation only from drivly. Currently USA Only
      parameters:
      - description: tokenId for vehicle to get valuation
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - valuations
  /v2/vehicles/{tokenId}/valuations:
    get:
      description: gets valuations for a particular user device. Includes only price
        valuations, not offers. gets list of most recent
      parameters:
      - description: tokenId for vehicle to get offers
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DIMO-Network_valuations-api_internal_core_models.DeviceValuation'
      security:
      - BearerAuth: []
      tags:
      - valuations
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"